Put separator-view in its own class (IOS-134)
aaaaaand improve A11y
